feat: add tables to AI Agent context
