+++ apps/sqlrooms-cli-ui/package.json
@@ -33,6 +33,7 @@
     "@sqlrooms/vega": "workspace:*",
     "@sqlrooms/webcontainer": "workspace:*",
     "@webcontainer/api": "^1.6.1",
+    "ai": "^6.0.154",
     "lucide-react": "^0.555.0",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
